requirements.txt
beautifulsoup4==4.9.1
bottle==0.12.18
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
gevent==20.5.2
greenlet==0.4.15
idna==2.9
pip==19.2.3
pycparser==2.20
setuptools==41.2.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
vkcoin==2.0.4
vkcoinapi==1.2.7
